+++ crypto_sign/dilithium2/ref/sign.h
@@ -14,23 +14,23 @@ void challenge(poly *c, const uint8_t seed[SEEDBYTES]);
 int crypto_sign_keypair(uint8_t *pk, uint8_t *sk);
 
 #define crypto_sign_signature DILITHIUM_NAMESPACE(signature)
-int crypto_sign_signature(uint8_t *sig, size_t *siglen,
-                          const uint8_t *m, size_t mlen,
+int crypto_sign_signature(uint8_t *sig, unsigned long long *siglen,
+                          const uint8_t *m, unsigned long long mlen,
                           const uint8_t *sk);
 
 #define crypto_sign DILITHIUM_NAMESPACETOP
-int crypto_sign(uint8_t *sm, size_t *smlen,
-                const uint8_t *m, size_t mlen,
+int crypto_sign(uint8_t *sm, unsigned long long *smlen,
+                const uint8_t *m, unsigned long long mlen,
                 const uint8_t *sk);
 
 #define crypto_sign_verify DILITHIUM_NAMESPACE(verify)
-int crypto_sign_verify(const uint8_t *sig, size_t siglen,
-                       const uint8_t *m, size_t mlen,
+int crypto_sign_verify(const uint8_t *sig, unsigned long long siglen,
+                       const uint8_t *m, unsigned long long mlen,
                        const uint8_t *pk);
 
 #define crypto_sign_open DILITHIUM_NAMESPACE(open)
-int crypto_sign_open(uint8_t *m, size_t *mlen,
-                     const uint8_t *sm, size_t smlen,
+int crypto_sign_open(uint8_t *m, unsigned long long *mlen,
+                     const uint8_t *sm, unsigned long long smlen,
                      const uint8_t *pk);
 
 #endif
